Revert "Add constants for class module"
This reverts commit 1254830f1bb4dd887467a8acefa9c85f45d72192.
