--- src/class/constants/index.js
@@ -1,19 +1,0 @@
-export { INPAGE_EVENT, INPAGE_ID_PREFIX } from './InpageProvider';
-
-export const REQUEST_TIMEOUT_MSEC = 30000;
-
-export const STORAGE_USER_META_KEY = 'endpass-user-meta';
-
-export const AVAILABLE_USER_META_PROPS = ['activeAccount'];
-
-export const PROXY_REQUEST_PREFIX = 'endpass-identity';
-
-export const HARDWARE_DERIVIATION_PATH = `m/44'/60'/0'/0/`;
-
-export default {
-  REQUEST_TIMEOUT_MSEC,
-  STORAGE_USER_META_KEY,
-  AVAILABLE_USER_META_PROPS,
-  PROXY_REQUEST_PREFIX,
-  HARDWARE_DERIVIATION_PATH,
-};
