+++ src/class/singleton/dappBridge.js
@@ -1,5 +1,5 @@
-import { INPAGE_EVENT } from '@/class/constants';
-import { EventEmitter } from '@/class/internal';
+import { INPAGE_EVENT } from '@/constants';
+import EventEmitter from '../EventEmitter';
 
 class DappBridge extends EventEmitter {
   constructor(...props) {
