+++ src/class/singleton/http.js
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
-import { REQUEST_TIMEOUT_MSEC } from '@/class/constants';
+import { REQUEST_TIMEOUT_MSEC } from '@/constants';
 
 const config = {
   timeout: REQUEST_TIMEOUT_MSEC,
