+++ src/class/storage/SettingsStorage.js
@@ -1,8 +1,5 @@
 import { pick } from 'lodash';
-import {
-  STORAGE_USER_META_KEY,
-  AVAILABLE_USER_META_PROPS,
-} from '@/class/constants';
+import { STORAGE_USER_META_KEY, AVAILABLE_USER_META_PROPS } from '@/constants';
 
 class SettingsStorage {
   constructor({ storage }) {
