+++ src/class/wallet/proxy/hardware/LedgerProxy.js
@@ -2,8 +2,9 @@ import Tx from 'ethereumjs-tx';
 import HDKey from 'ethereumjs-wallet/hdkey';
 import LedgerTransport from '@ledgerhq/hw-transport-u2f'; // for browser
 import Eth from '@ledgerhq/hw-app-eth';
-import { NotificationError, web3 } from '@/class/internal';
-import { HARDWARE_DERIVIATION_PATH } from '@/class/constants';
+import { NotificationError } from '@/class/internal';
+import web3 from '@/class/singleton/web3';
+import { HARDWARE_DERIVIATION_PATH } from '@/constants';
 import getChildrenAddress from '../utils/getChildrenAddress';
 
 const { sha3, toHex } = web3.utils;
