+++ src/constants/index.js
@@ -1,3 +1,5 @@
+export { INPAGE_EVENT, INPAGE_ID_PREFIX } from './InpageProvider';
+
 export const MAIN_NET_ID = 1;
 
 export const NETWORK_URL = Object.freeze({
@@ -76,6 +78,16 @@ export const IDENTITY_MODE = Object.freeze({
   LOCAL: 'local',
 });
 
+export const PROXY_REQUEST_PREFIX = 'endpass-identity';
+
+export const STORAGE_USER_META_KEY = 'endpass-user-meta';
+
+export const AVAILABLE_USER_META_PROPS = ['activeAccount'];
+
+export const REQUEST_TIMEOUT_MSEC = 30000;
+
+export const HARDWARE_DERIVIATION_PATH = `m/44'/60'/0'/0/`;
+
 export const HARDWARE_WALLET_TYPE = Object.freeze({
   TREZOR: 'TrezorAccount',
   LEDGER: 'LedgerAccount',
@@ -109,6 +121,11 @@ export default {
   DEFAULT_NETWORKS,
   CURRENCIES,
   IDENTITY_MODE,
+  PROXY_REQUEST_PREFIX,
+  STORAGE_USER_META_KEY,
+  AVAILABLE_USER_META_PROPS,
+  REQUEST_TIMEOUT_MSEC,
+  HARDWARE_DERIVIATION_PATH,
   HARDWARE_WALLET_TYPE,
   WALLET_TYPE,
   DAPP_WHITELISTED_METHODS,
