+++ test/unit/specs/services/localSettings.spec.js
@@ -1,5 +1,5 @@
 import { LocalStorage } from '@/class';
-import { STORAGE_USER_META_KEY } from '@/class/constants';
+import { STORAGE_USER_META_KEY } from '@/constants';
 
 const localSettings = require.requireActual('@/services/localSettings').default;
